Move `RunConfigSettings` in the same package as `RunConfig`
