+++ src/main/java/net/fabricmc/loom/configuration/ide/SetupIntelijRunConfigs.java
@@ -32,7 +32,6 @@ import org.apache.commons.io.FileUtils;
 import org.gradle.api.Project;
 
 import net.fabricmc.loom.LoomGradleExtension;
-import net.fabricmc.loom.configuration.RunConfigSettings;
 import net.fabricmc.loom.configuration.providers.minecraft.MinecraftNativesProvider;
 import net.fabricmc.loom.configuration.providers.minecraft.assets.MinecraftAssetsProvider;
