+++ src/main/java/net/fabricmc/loom/task/GenEclipseRunsTask.java
@@ -33,8 +33,8 @@ import org.gradle.api.tasks.TaskAction;
 import org.gradle.plugins.ide.eclipse.model.EclipseModel;
 
 import net.fabricmc.loom.LoomGradleExtension;
-import net.fabricmc.loom.configuration.RunConfigSettings;
 import net.fabricmc.loom.configuration.ide.RunConfig;
+import net.fabricmc.loom.configuration.ide.RunConfigSettings;
 
 public class GenEclipseRunsTask extends AbstractLoomTask {
 	@TaskAction
