+++ src/main/java/net/fabricmc/loom/task/GenIdeaProjectTask.java
@@ -45,8 +45,8 @@ import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
 import net.fabricmc.loom.LoomGradleExtension;
-import net.fabricmc.loom.configuration.RunConfigSettings;
 import net.fabricmc.loom.configuration.ide.RunConfig;
+import net.fabricmc.loom.configuration.ide.RunConfigSettings;
 
 public class GenIdeaProjectTask extends AbstractLoomTask {
 	@TaskAction
